+++ r3dg-rasterization/cuda_rasterizer/rasterizer.h
@@ -44,7 +44,7 @@ namespace CudaRasterizer
 			const float scale_modifier,
 			const float* rotations,
 			const float* cov3D_precomp,
-			const int* shaderAddresses,
+			const int64_t* shaderAddresses,
 			const float* viewmatrix,
 			const float* viewmatrix_inv,
 			const float* projmatrix,
